requirements.txt
psutil>=5.9.0
google-generativeai>=0.3.0
